requirements.txt
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
SQLAlchemy==1.3.12
visitor==0.1.3
Werkzeug==0.16.0
flask-cors==3.0.8